cleanup(core): split the performance analysis into its own module
Move PerformanceAnalysis and its helpers into performance-analysis.ts so the
lifecycle file is just the collector, drop the never-displayed finish-chain
diagnostic, and route formatDuration through the shared native formatter.
